+++ esp-hal/src/soc/esp32c2/mod.rs
@@ -28,9 +28,9 @@ pub(crate) mod registers {
 
 pub(crate) mod constants {
     /// The lower bound of the system's DRAM (Data RAM) address space.
-    pub const SOC_DRAM_LOW: u32 = 0x3FCA_0000;
+    pub const SOC_DRAM_LOW: usize = 0x3FCA_0000;
     /// The upper bound of the system's DRAM (Data RAM) address space.
-    pub const SOC_DRAM_HIGH: u32 = 0x3FCE_0000;
+    pub const SOC_DRAM_HIGH: usize = 0x3FCE_0000;
 
     /// RC FAST Clock value (Hertz).
     pub const RC_FAST_CLK: fugit::HertzU32 = fugit::HertzU32::kHz(17500);
